Move Java demos to com.example package (#412)
