+++ .vscode/cspell.json
@@ -4,6 +4,7 @@
     "words": [
         "Blobject",
         "icegrid",
+        "visitorcenter",
         "zeroc"
     ],
     "dictionaries": [
