+++ java/Ice/greeter/serveramd/src/main/java/com/example/ice/greeter/serveramd/Chatbot.java
@@ -1,8 +1,8 @@
 // Copyright (c) ZeroC, Inc.
 
-package com.zeroc.demos.ice.greeter.serveramd;
+package com.example.ice.greeter.serveramd;
 
-import com.zeroc.demos.visitor_center.Greeter;
+import com.example.visitorcenter.Greeter;
 import com.zeroc.Ice.Current;
 import java.util.concurrent.CompletableFuture;
 import java.util.concurrent.CompletionStage;
